pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "news_agent"
version = "0.1.0"
description = "Automotive / economy news aggregator — Stage 1 MVP"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "NewsMaker" }]
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    "feedparser>=6.0",
    "trafilatura>=1.12",
    "readability-lxml>=0.8.1",
    "rapidfuzz>=3.9",
    "python-dotenv>=1.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "PyYAML>=6.0",
    "google-api-python-client>=2.130",
    "google-auth>=2.29",
    "google-auth-httplib2>=0.2",
    "anthropic>=0.39",
    "openai>=1.40",
    "tenacity>=8.3",
    "structlog>=24.1",
    "typer>=0.12",
    "rich>=13.7",
    "tiktoken>=0.7",
]

[project.optional-dependencies]
js = ["playwright>=1.44"]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
    "ruff>=0.5",
    "black>=24.4",
    "mypy>=1.10",
    "types-PyYAML",
    "types-beautifulsoup4",
]

[project.scripts]
news_agent = "news_agent.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/news_agent"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM", "PL", "RUF"]
ignore = ["PLR0913", "PLR0912", "PLR0915", "PLR2004", "E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"
explicit_package_bases = true
disallow_untyped_decorators = false
# Third-party packages without stubs
[[tool.mypy.overrides]]
module = [
    "feedparser",
    "trafilatura",
    "readability",
    "readability.*",
    "googleapiclient",
    "googleapiclient.*",
    "google.oauth2",
    "google.oauth2.*",
    "google_auth_httplib2",
    "rapidfuzz",
    "rapidfuzz.*",
    "tiktoken",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers"
